fix: add stories and refactor the member stats screen
