+++ src/modules/builder/map/components/map/ItemsMarkers.tsx
@@ -4,8 +4,6 @@ import MarkerClusterGroup from 'react-leaflet-cluster';
 
 import { Box, CircularProgress, Stack } from '@mui/material';
 
-import { useItemsInMap } from '@/query/hooks/itemGeolocation';
-
 import { useQueryClientContext } from '../context/QueryClientContext';
 import ItemMarker from './ItemMarker';
 
@@ -21,7 +19,7 @@ const ItemsMarkers = ({
     lng2: number;
   };
 }): JSX.Element | JSX.Element[] | undefined => {
-  const { item } = useQueryClientContext();
+  const { item, useItemsInMap } = useQueryClientContext();
   const { data: itemGeolocations, isFetching } = useItemsInMap({
     ...bounds,
     parentItemId: item?.id,
