+++ src/modules/home/memberAnalytics/MemberStatsCard.stories.tsx
@@ -1,18 +1,17 @@
 import { Meta, StoryObj } from '@storybook/react';
 
-import { Image } from './StyledImages';
+import { MemberStatsCard } from './MemberStatsCard';
 
 const meta = {
-  component: Image,
-} satisfies Meta<typeof Image>;
+  component: MemberStatsCard,
+} satisfies Meta<typeof MemberStatsCard>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
 export const Default = {
   args: {
-    src: '/illustration/teacher-red.webp',
-    width: 300,
-    height: 300,
+    title: 'My great stat',
+    stat: 120,
   },
 } satisfies Story;
